requirements.txt
mysql-connector-python>=8.0.13
Click>=7.0
GitPython>=2.1.11
gitdb2>=2.0.5
smmap2>=2.0.5
